+++ app/product/infra/kafka/model/delete_product.go
@@ -1,5 +1,5 @@
 package model
 
 type DeleteProductSendToKafka struct {
-	ID int64 `json:"id,omitempty"`
+	ID int32 `json:"id,omitempty"`
 }
